# runner-dashboard/openapi/runner-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Playwright Automation Runner API",
    "description": "REST control plane for launching and observing allowlisted automation suites. The SSE event stream (GET /api/v1/runs/{runId}/events) is intentionally not part of this document - see docs/SSE_CONTRACT_V1.md.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/v1/runs": {
      "get": {
        "tags": [
          "run-controller"
        ],
        "operationId": "listRuns",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "run-controller"
        ],
        "operationId": "createRun",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Run accepted and queued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request, or an environment/suite combination not allowlisted.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "503": {
            "description": "The run queue is full, the runner is degraded, or the canonical event journal is unavailable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{runId}/cancel": {
      "post": {
        "tags": [
          "run-controller"
        ],
        "operationId": "cancelRun",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cancellation requested (or the run was already terminal, a no-op).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "404": {
            "description": "No run exists for the given runId.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "503": {
            "description": "The runner is degraded or the canonical event journal is unavailable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tests": {
      "get": {
        "tags": [
          "test-catalog-controller"
        ],
        "operationId": "listPublicTests",
        "parameters": [
          {
            "name": "environment",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "PUBLIC",
                "LOCAL"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestCatalogResponse"
                }
              }
            }
          },
          "400": {
            "description": "environment is missing or not PUBLIC.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "503": {
            "description": "The committed test catalog file is missing or unreadable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{runId}": {
      "get": {
        "tags": [
          "run-controller"
        ],
        "operationId": "getRun",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "404": {
            "description": "No run exists for the given runId.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{runId}/log": {
      "get": {
        "tags": [
          "run-controller"
        ],
        "operationId": "downloadRunLog",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "No run exists for the given runId, or its process log does not exist yet.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{runId}/artifacts": {
      "get": {
        "tags": [
          "artifact-controller"
        ],
        "operationId": "listRunArtifacts",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "testId",
            "in": "query",
            "description": "Narrows the result to one test's own artifacts, if given.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArtifactSummaryResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No run exists for the given runId.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error, or the run's artifact manifest is corrupt.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{runId}/artifacts/{artifactId}": {
      "get": {
        "tags": [
          "artifact-controller"
        ],
        "operationId": "downloadRunArtifact",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "artifactId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "application/zip": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "video/webm": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "No run exists for the given runId, or no artifact exists for the given artifactId.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error, or the run's artifact manifest is corrupt.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/capabilities": {
      "get": {
        "tags": [
          "capabilities-controller"
        ],
        "operationId": "getRunnerCapabilities",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilitiesResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateRunRequest": {
        "type": "object",
        "properties": {
          "environment": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "LOCAL"
            ]
          },
          "suite": {
            "type": "string",
            "enum": [
              "SMOKE",
              "API",
              "UI",
              "JOURNEY",
              "REGRESSION",
              "FIXTURE",
              "CUSTOM"
            ]
          },
          "testKeys": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "environment",
          "suite"
        ]
      },
      "RunResponse": {
        "type": "object",
        "properties": {
          "runId": {
            "type": "string"
          },
          "environment": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "LOCAL"
            ]
          },
          "suite": {
            "type": "string",
            "enum": [
              "SMOKE",
              "API",
              "UI",
              "JOURNEY",
              "REGRESSION",
              "FIXTURE",
              "CUSTOM"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "QUEUED",
              "STARTING",
              "RUNNING",
              "SUCCEEDED",
              "FAILED",
              "CANCELLED",
              "TIMED_OUT",
              "ERROR"
            ]
          },
          "requestedAt": {
            "type": "string",
            "format": "date-time"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "exitCode": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": "string"
          },
          "processLogUrl": {
            "type": "string"
          },
          "selectedTests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SelectedTestResponse"
            }
          }
        },
        "required": [
          "environment",
          "processLogUrl",
          "requestedAt",
          "runId",
          "selectedTests",
          "status",
          "suite"
        ]
      },
      "SelectedTestResponse": {
        "type": "object",
        "properties": {
          "testKey": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "layer": {
            "type": "string",
            "enum": [
              "API",
              "UI",
              "JOURNEY"
            ]
          }
        },
        "required": [
          "displayName",
          "layer",
          "testKey"
        ]
      },
      "ProblemDetail": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "properties": {
            "type": "object",
            "additionalProperties": {}
          }
        },
        "required": [
          "detail",
          "instance",
          "status",
          "title"
        ]
      },
      "TestCatalogEntry": {
        "type": "object",
        "properties": {
          "testKey": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "API",
              "UI",
              "JOURNEY"
            ]
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          }
        },
        "required": [
          "category",
          "displayName",
          "tags",
          "testKey"
        ]
      },
      "TestCatalogResponse": {
        "type": "object",
        "properties": {
          "tests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TestCatalogEntry"
            }
          }
        },
        "required": [
          "tests"
        ]
      },
      "ArtifactSummaryResponse": {
        "type": "object",
        "properties": {
          "artifactId": {
            "type": "string"
          },
          "testId": {
            "type": "string"
          },
          "testDisplayName": {
            "type": "string"
          },
          "stepId": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "SCREENSHOT",
              "TRACE",
              "VIDEO"
            ]
          },
          "mediaType": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "downloadUrl": {
            "type": "string"
          }
        },
        "required": [
          "artifactId",
          "createdAt",
          "downloadUrl",
          "mediaType",
          "sizeBytes",
          "testDisplayName",
          "testId",
          "type"
        ]
      },
      "CapabilitiesResponse": {
        "type": "object",
        "properties": {
          "apiVersion": {
            "type": "string"
          },
          "eventSchemaVersion": {
            "type": "string"
          },
          "environments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EnvironmentCapabilities"
            }
          }
        },
        "required": [
          "apiVersion",
          "environments",
          "eventSchemaVersion"
        ]
      },
      "EnvironmentCapabilities": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "LOCAL"
            ]
          },
          "suites": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "SMOKE",
                "API",
                "UI",
                "JOURNEY",
                "REGRESSION",
                "FIXTURE",
                "CUSTOM"
              ]
            }
          }
        },
        "required": [
          "name",
          "suites"
        ]
      }
    }
  }
}